TASK-7 | feat: refactored constants and imports
