+++ src/constants/user-data.js
@@ -1,0 +1,8 @@
+export const userData = {
+  cardNumber: '4242 4242 4242 4242',
+  cardExpire: '04 / 24',
+  cardCVC: '000',
+  firstName: 'First name',
+  lastName: 'Last name',
+  email: '',
+};
