+++ src/elements/index.js
@@ -1,0 +1,9 @@
+const { BaseElement } = require('./base-element');
+const { ButtonElement } = require('./button-element');
+const { CheckboxElement } = require('./checkbox-element');
+const { InputElement } = require('./input-element');
+
+exports.BaseElement = BaseElement;
+exports.ButtonElement = ButtonElement;
+exports.CheckboxElement = CheckboxElement;
+exports.InputElement = InputElement;
