+++ tests/example.spec.js
@@ -1,13 +1,6 @@
 const { test } = require('../src/fixtures/fixtures');
+const { userData } = require('../src/constants/user-data');
 
-const userData = {
-  cardNumber: '4242 4242 4242 4242',
-  cardExpire: '04 / 24',
-  cardCVC: '000',
-  firstName: 'First name',
-  lastName: 'Last name',
-  email: '',
-};
 test.beforeEach(async ({ basePage }) => {
   await basePage.visit('/qa-test-7R58U3/');
 });
